Dissolve dependency chains among benches

+++ benches/f32_cbrt.rs
@@ -1,9 +1,9 @@
 mod bench;
 
-bench!(bench_metallic, metallic::f64::cbrt);
-bench!(bench_core_math, core_math::cbrt);
-bench!(bench_std, f64::cbrt);
-bench!(bench_libm, libm::cbrt);
+bench!(bench_metallic, metallic::f32::cbrt);
+bench!(bench_core_math, core_math::cbrtf);
+bench!(bench_std, f32::cbrt);
+bench!(bench_libm, libm::cbrtf);
 
 criterion::criterion_group!(
     benches,
